backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.6.0
sqlalchemy>=2.0.25
pandas>=2.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
python-dotenv>=1.0.0
openai>=1.45.0
sqlparse